Added fallback support.
A regex pattern can be specified in the proxy.yml config now. Whenever possible (all post-authentication events where a server has been selected), if the user is disconnected and the disconnect message matches the specified kickRegex message, then it will try to find a backup server to send them to based on the servers list in routes (usually a hub server).

This solution is pretty messy as it requires variadic arguments (in place of passing in null since that doesn't exist as far as my limited understanding of Go would say) but it works with minimal code changes.